Refactor EV data parsing
EVs were previously saved as a string, but I re-factored them to
be stored as an object instead. Now they are parsed to a string @
export time by a function called createStatString.

It is written in such a way that it is not a problem for any number
of the stats to be undefined. It is still a completely optional
dimension.