Let user write over initial value for species
Because the inputs had their values set to ALWAYS = pokeObj.prop,
the user was not able to overwrite these values.

With the use of useEffect hooks and querySelector('#species')
within each component grabbed by their unique slotIDs, I simply
set the .value of the species input once when the pokeObj is
first sent to the component.