TeamSlot and StatInputs update in-synch
TeamSlot and StatInput components now update synchronously when a
new pokemon object is generated and sent to the TeamSlot level.
A signal in the form of a date state is updated in the root
component when a new mon is generated. useEffect hook in the
TeamSlot component listens for this change to trigger update in
gui.

Stats are sent to StatInput components via useContext. OnChange,
they are updated in the temporary obj inside their TeamSlot
parent component. None of these components yet have the
functionality to change state in the root component.