Changing inputs updates local obj state
The local, temporary pokemon obj in the TeamSlot component state
now is modified to the current user input onChange whenever the
user makes changes to their team slot.

However, this is not yet updated in the app's main state. This is
by design. I don't want the user's data being sent to where it
can be exported until the user says they are ready.